final backend code

+++ src/main/java/com/capstone/mapapi/destination/DestinationRepository.java
@@ -1,8 +1,7 @@
 package com.capstone.mapapi.destination;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.List;
 
-public interface DestinationRepository extends JpaRepository<Destination, Long> {
+public interface DestinationRepository extends JpaRepository<Destination, Integer> {
     Destination findTopByUserIDOrderByDestinationIdDesc(String userID);
 }
